global MagicLinkInstance
