+++ src/lib/session/Provider.tsx
@@ -1,4 +1,7 @@
-import { createContext, useReducer, ReactNode } from 'react'
+import { createContext, useReducer, ReactNode, useEffect, useState } from 'react'
+import { Magic } from 'magic-sdk'
+import { OAuthExtension } from '@magic-ext/oauth'
+import { assert, assertInt } from '@lib'
 import { SessionState, initialState } from './state'
 import { Dispatch } from './actions'
 import { reducer } from './reducer'
@@ -15,9 +18,24 @@ interface ProviderProps {
 export const SessionProvider = ({ children }: ProviderProps) => {
   const [state, dispatch] = useReducer(reducer, initialState)
 
+  const [magic, setMagic] = useState<MagicLinkInstance>()
+
+  useEffect(() => {
+    const rpcUrl = assert(process.env.NEXT_PUBLIC_RPC_URL, 'NEXT_PUBLIC_RPC_URL')
+    const chainId = assertInt(process.env.NEXT_PUBLIC_CHAIN_ID, 'NEXT_PUBLIC_CHAIN_ID')
+    const magicPublicKey = assert(process.env.NEXT_PUBLIC_MAGIC_PUBLISHABLE_KEY, 'NEXT_PUBLIC_MAGIC_PUBLISHABLE_KEY')
+    const newMagic = new Magic(magicPublicKey, {
+      network: { rpcUrl, chainId },
+      extensions: [new OAuthExtension()],
+    })
+    setMagic(newMagic)
+  }, [])
+
   return (
     <SessionContext.Provider value={state}>
-      <SessionDispatchContext.Provider value={dispatchMiddleware(dispatch)}>{children}</SessionDispatchContext.Provider>
+      <SessionDispatchContext.Provider value={dispatchMiddleware(dispatch, magic)}>
+        {children}
+      </SessionDispatchContext.Provider>
     </SessionContext.Provider>
   )
 }
