+++ src/types/global.d.ts
@@ -26,6 +26,8 @@ declare global {
   type NextJsInitializedPage = (props: AppProps) => JSX.Element
 
   export type CheckoutMethodType = 'usd' | 'polygon' | 'ethereum'
+
+  type MagicLinkInstance = InstanceWithExtensions<SDKBase, OAuthExtension[]>
 }
 
 export {}
